Business Plan Checker/requirements.txt
streamlit
PyPDF2
google-genai
python-dotenv
